Slate/UMG - Moving the SConstraintCanvas out of the UMG Module and into Slate proper.  This will allow usage in more core engine modules that come before UMG.
[CL 2598074 by Nick Darnell in Main branch]

+++ Engine/Source/Runtime/Slate/Public/Widgets/Layout/SConstraintCanvas.cpp
@@ -1,6 +1,8 @@
 // Copyright 1998-2015 Epic Games, Inc. All Rights Reserved.
 
-#include "UMGPrivatePCH.h"
+#include "SlatePrivatePCH.h"
+
+#include "SConstraintCanvas.h"
 #include "LayoutUtils.h"
 
 
@@ -10,7 +12,6 @@
 SConstraintCanvas::SConstraintCanvas()
 : Children()
 {
-
 }
 
 void SConstraintCanvas::Construct( const SConstraintCanvas::FArguments& InArgs )
@@ -44,7 +45,7 @@ int32 SConstraintCanvas::RemoveSlot( const TSharedRef<SWidget>& SlotWidget )
 struct FChildZOrder
 {
 	int32 ChildIndex;
-	int32 ZOrder;
+	float ZOrder;
 };
 
 struct FSortSlotsByZOrder
